+++ libpm/galgas2/AC_GALGAS_uniqueMap.cpp
@@ -1495,7 +1495,7 @@ typeComparisonResult cSharedUniqueMapRoot::mapCompare (const cSharedUniqueMapRoo
   }else{
     capCollectionElementArray array ; populateEnumerationArray (array, kENUMERATION_UP) ;
     capCollectionElementArray operandArray ; inOperand->populateEnumerationArray (operandArray, kENUMERATION_UP) ;
-    for (uint32_t i=0 ; (i<array.count ()) && (kOperandEqual == result) ; i++) {
+    for (int32_t i=0 ; (i<array.count ()) && (kOperandEqual == result) ; i++) {
       result = array.objectAtIndex (i COMMA_HERE).compare (operandArray.objectAtIndex (i COMMA_HERE)) ;
     }
     if (kOperandEqual == result) {
@@ -2017,7 +2017,7 @@ static void enterDescendingEnumeration (cUniqueMapNode * inNode,
 void cSharedUniqueMapRoot::populateEnumerationArray (capCollectionElementArray & ioEnumerationArray,
                                                      const typeEnumerationOrder inEnumerationOrder) const {
   // printf ("MAP COUNT %u\n", count ()) ;
-  ioEnumerationArray.setCapacity (mNodeCount) ;
+  ioEnumerationArray.setCapacity ((int32_t) mNodeCount) ;
   switch (inEnumerationOrder) {
   case kENUMERATION_UP  :
     enterAscendingEnumeration (mRoot, ioEnumerationArray) ;
@@ -2026,7 +2026,9 @@ void cSharedUniqueMapRoot::populateEnumerationArray (capCollectionElementArray &
     enterDescendingEnumeration (mRoot, ioEnumerationArray) ;
     break ;
   }
-  MF_Assert (mNodeCount == ioEnumerationArray.count (), "mNodeCount (%lld) != ioEnumerationArray.count () (%lld)", mNodeCount, ioEnumerationArray.count ()) ;
+  MF_Assert (mNodeCount == (uint32_t) ioEnumerationArray.count (),
+             "mNodeCount (%lld) != ioEnumerationArray.count () (%lld)",
+             mNodeCount, ioEnumerationArray.count ()) ;
 }
 
 //---------------------------------------------------------------------------------------------------------------------*
